Implemented first cut of a menu interface and menu stack.
Pressing Enter in the main menu switches to the choose animation menu.
Pressing Up/Down in the choose animation menu switches (instantly) to
the next animation on the SD card (no api exists yet for selecting the
previous animation on the SD card).
Pressing Back in any menu exits the menu and restores the previous menu.